fix(navigation): add NavigationResult type type NavigationResult = PipelineResult | boolean; ref: https://discourse.aurelia.io/t/checking-for-candeactivate-before-navigatetoroute/51/2
