+++ src/router.js
@@ -8,7 +8,7 @@ import {
   _normalizeAbsolutePath,
   _createRootedPath,
   _resolveUrl} from './util';
-import {RouteConfig} from './interfaces';
+import {RouteConfig, NavigationResult} from './interfaces';
 
 /**
 * The primary class responsible for handling routing and navigation.
@@ -165,7 +165,7 @@ export class Router {
   * @param fragment The URL fragment to use as the navigation destination.
   * @param options The navigation options.
   */
-  navigate(fragment: string, options?: any): boolean {
+  navigate(fragment: string, options?: any): NavigationResult {
     if (!this.isConfigured && this.parent) {
       return this.parent.navigate(fragment, options);
     }
@@ -182,7 +182,7 @@ export class Router {
   * @param params The route parameters to be used when populating the route pattern.
   * @param options The navigation options.
   */
-  navigateToRoute(route: string, params?: any, options?: any): boolean {
+  navigateToRoute(route: string, params?: any, options?: any): NavigationResult {
     let path = this.generate(route, params);
     return this.navigate(path, options);
   }
